超导把数据处理放在model中
1 超导把数据处理放在model中
2  SMDetailViewController.m 登录切换后实现试听与正试播放切换

+++ myoffer/MyOffer/Class/ApplyList/controller/ApplyViewController.m
@@ -879,7 +879,7 @@ typedef NS_ENUM(NSInteger,ApplyTableStatus){
     
     NSInteger count = self.cancelSetions.count + self.cancelindexPathes.count;
     
-    self.cancelCountLab.text = [NSString stringWithFormat:@"已选择 : %ld ", count];
+    self.cancelCountLab.text = [NSString stringWithFormat:@"已选择 : %ld ", (long)count];
     
     self.cancelBottomButton.enabled = count;
